Cambia tipo de variable por una de menor tamaño

+++ projectFunctions.c
@@ -215,7 +215,7 @@ void imprimeMenu( void )
 /*******************************************************************************************/
 /*******************************************************************************************/
 
-long unsigned int valorDeRetardo( int velSecuencia )
+int valorDeRetardo( int velSecuencia )
 {
   switch( velSecuencia )
   {
